refactored Task Label and Due Date
